refactor: simplificar tipos e testes da abertura linha quente
Unifica DecisaoAbertura, evita segundo estadoEmJogo na abertura,
colapsa testes redundantes e valida ausência de sorteio no teste do orchestrator.
